feat: empty state for earn dashboard (#4133)
